requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiodns==2.0.0             # via nubosas.aiotemplate (setup.py)
aiohttp==3.6.2            # via nubosas.aiotemplate (setup.py)
aiopg==1.0.0              # via nubosas.aiotemplate (setup.py)
alembic==1.4.2            # via nubosas.aiotemplate (setup.py)
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp
brotlipy==0.7.0           # via nubosas.aiotemplate (setup.py)
cchardet==2.1.6           # via nubosas.aiotemplate (setup.py)
cffi==1.14.0              # via brotlipy, pycares
chardet==3.0.4            # via aiohttp, nubosas.aiotemplate (setup.py)
idna==2.9                 # via yarl
mako==1.1.2               # via alembic
markupsafe==1.1.1         # via mako
multidict==4.7.5          # via aiohttp, yarl
psycopg2-binary==2.8.5    # via aiopg
pycares==3.1.1            # via aiodns
pycparser==2.20           # via cffi
python-dateutil==2.8.1    # via alembic
python-editor==1.0.4      # via alembic
six==1.14.0               # via python-dateutil, sqlalchemy-utils
sqlalchemy-utils==0.36.4  # via nubosas.aiotemplate (setup.py)
sqlalchemy==1.3.16        # via alembic, nubosas.aiotemplate (setup.py), sqlalchemy-utils
yarl==1.4.2               # via aiohttp
